requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
bcrypt==4.0.1
argon2-cffi==21.3.0
passlib==1.7.4
tqdm==4.65.0